--- src/pages/HomePage.tsx
@@ -1,6 +1,0 @@
-import HomePageContainer from '@/components/home/containers/HomePageContainer';
-
-export default function HomePage() {
-  return <HomePageContainer />;
-}
-
