--- src/pages/diary/DiaryEditPage.tsx
@@ -1,6 +1,0 @@
-import DiaryEditPageContainer from '@/components/diary/containers/DiaryEditPageContainer';
-
-export default function DiaryEditPage() {
-  return <DiaryEditPageContainer />;
-}
-
