--- src/pages/diary/DiaryPage.tsx
@@ -1,5 +1,0 @@
-import DiaryPageContainer from '@/components/diary/containers/DiaryPageContainer';
-
-export default function DiaryPage() {
-  return <DiaryPageContainer />;
-}
