+++ src/pages/diary/diaryeditpage/index.tsx
@@ -1,0 +1,5 @@
+import DiaryEditPageContainer from '@/pages/diary/diaryeditpage/DiaryEditPageContainer';
+
+export default function DiaryEditPage() {
+  return <DiaryEditPageContainer />;
+}
