+++ src/pages/diary/diarypage/index.tsx
@@ -1,0 +1,5 @@
+import DiaryPageContainer from '@/pages/diary/diarypage/DiaryPageContainer';
+
+export default function DiaryPage() {
+  return <DiaryPageContainer />;
+}
