+++ src/pages/home/index.tsx
@@ -1,0 +1,5 @@
+import HomePageContainer from '@/pages/home/HomePageContainer';
+
+export default function HomePage() {
+  return <HomePageContainer />;
+}
