--- src/pages/mypage/MyEditPage.tsx
@@ -1,5 +1,0 @@
-import MyEditPageContainer from "@/components/mypage/containers/MyEditPageContainer";
-
-export default function MyEditPage() {
-  return <MyEditPageContainer />;
-}
