--- src/pages/mypage/MyPage.tsx
@@ -1,5 +1,0 @@
-import MyPageContainer from '@/components/mypage/containers/MyPageContainer';
-
-export default function MyPage() {
-  return <MyPageContainer />;
-}
