+++ src/pages/mypage/myeditpage/index.tsx
@@ -1,0 +1,5 @@
+import MyEditPageContainer from "@/pages/mypage/myeditpage/MyEditPageContainer";
+
+export default function MyEditPage() {
+  return <MyEditPageContainer />;
+}
