+++ src/pages/mypage/mypage/index.tsx
@@ -1,0 +1,5 @@
+import MyPageContainer from '@/pages/mypage/mypage/MyPageContainer';
+
+export default function MyPage() {
+  return <MyPageContainer />;
+}
