--- src/pages/report/ReportPage.tsx
@@ -1,6 +1,0 @@
-import ReportPageContainer from "@/components/report/containers/ReportPageContainer";
-
-export default function ReportPage() {
-  return <ReportPageContainer />;
-}
-
