+++ src/pages/report/index.tsx
@@ -1,0 +1,6 @@
+import ReportPageContainer from "@/pages/report/ReportPageContainer";
+
+export default function ReportPage() {
+  return <ReportPageContainer />;
+}
+
